pyproject.toml
[project]
name = "eventsource-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.0.0",
    "logicblocks-event-store>=0.1.3",
    "pydantic>=2.10.6",
    "pytest>=8.3.5",
    "ruff>=0.9.10",
]
